chore: formating and removing unnecessary lines
