+++ src/rclient/cli.go
@@ -1,0 +1,105 @@
+package rclient
+
+import (
+	"encoding/json"
+	"fmt"
+	"strings"
+)
+
+// ShowMetrics parse metrics from Http WEB app show it out.
+func (rc *RClient) ShowMetrics() error {
+
+	if rc.SOAP == nil {
+		err := fmt.Errorf("Client SOA is not initializated")
+		return err
+	}
+
+	if rc.Metrics == nil {
+		err := fmt.Errorf("Metrics was not initializaded. Use UpdateMetrics()")
+		return err
+	}
+
+	for k, v := range rc.Metrics.Counters {
+		var d DataInMetricCount
+		metricName := strings.Replace(k, ".", "_", -1)
+		if !strings.HasPrefix(metricName, "rundeck") {
+			continue
+		}
+
+		b, e := json.Marshal(v)
+		if e != nil {
+			fmt.Println(e)
+		}
+
+		e = json.Unmarshal(b, &d)
+		if e != nil {
+			fmt.Println(e)
+		}
+
+		fmt.Println(metricName+"_total : ", d.Count)
+	}
+
+	for k, v := range rc.Metrics.Gauges {
+		var d DataInMetricGauges
+		metricName := strings.Replace(k, ".", "_", -1)
+		if !strings.HasPrefix(metricName, "rundeck") {
+			metricName = "rundeck_" + metricName
+		}
+
+		b, e := json.Marshal(v)
+		if e != nil {
+			fmt.Println(e)
+		}
+
+		e = json.Unmarshal(b, &d)
+		if e != nil {
+			fmt.Println(e)
+		}
+
+		fmt.Println(metricName+"_total : ", d.Value)
+	}
+
+	for k, v := range rc.Metrics.Meters {
+		var d DataInMetricMeters
+		metricName := strings.Replace(k, ".", "_", -1)
+		if !strings.HasPrefix(metricName, "rundeck") {
+			metricName = "rundeck_servlet_" + metricName
+		}
+
+		b, e := json.Marshal(v)
+		if e != nil {
+			fmt.Println(e)
+		}
+
+		e = json.Unmarshal(b, &d)
+		if e != nil {
+			fmt.Println(e)
+		}
+
+		fmt.Println(metricName+"_total : ", d.Count)
+		fmt.Println(metricName+"_rate : ", d.M1Rate, d.M5Rate, d.M15Rate, d.MeanRate)
+	}
+
+	for k, v := range rc.Metrics.Timers {
+		var d DataInMetricTimers
+		metricName := strings.Replace(k, ".", "_", -1)
+		if !strings.HasPrefix(metricName, "rundeck") {
+			metricName = "rundeck_" + metricName
+		}
+
+		b, e := json.Marshal(v)
+		if e != nil {
+			fmt.Println(e)
+		}
+
+		e = json.Unmarshal(b, &d)
+		if e != nil {
+			fmt.Println(e)
+		}
+
+		fmt.Println(metricName+"_total : ", d.Count)
+		fmt.Println(metricName+"_rate : ", d.M1Rate, d.M5Rate, d.M15Rate, d.MeanRate)
+	}
+
+	return nil
+}
